nsysu_lab/myanti_platformapp/crossNoso/myanti_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-mail==1.5.8
h11==0.14.0
idna==3.10
jinja2>=3.1.6,<4.0.0
kaleido==0.2.1
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
pydantic==2.12.2
python-multipart==0.0.12
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.15.0
uvicorn==0.34.0
